backend/requirements-ref.txt
python-dotenv==1.0.1
tenacity!=8.4.0
openai==1.56.0
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.9
langchain-neo4j==0.1.1
langchain-huggingface
langchain-google-genai
neo4j==5.27.0
textblob==0.18.0.post0
nltk==3.8.1
sentence-transformers
google-generativeai
pymupdf
fastapi
uvicorn
flask
python-docx
pytest

# neo4j>=5.16
# sentence-transformers>=2.5
# openai>=1.13
# google-generativeai>=0.5
# fastapi
fastmcp
# uvicorn